+++ .scalafix.conf
@@ -1,7 +1,7 @@
 OrganizeImports {
   blankLines = Auto
   coalesceToWildcardImportThreshold = 5
-  groups = ["habla.", "org.apache.spark.", "*"]
+  groups = ["doric.", "org.apache.spark.", "*"]
   groupedImports = Merge
   importSelectorsOrder = Ascii
   removeUnused = true
