+++ core/src/main/scala/doric/syntax/BooleanColumns.scala
@@ -1,8 +1,8 @@
 package doric
 package syntax
 
-import cats.implicits._
 import doric.DoricColumn.sparkFunction
+
 import org.apache.spark.sql.{functions => f}
 
 private[syntax] trait BooleanColumns {
@@ -61,24 +61,5 @@ private[syntax] trait BooleanColumns {
       */
     def ||(other: DoricColumn[Boolean]): DoricColumn[Boolean] =
       or(other)
-
-    /**
-      * Returns null if the condition is true, and throws an exception otherwise.
-      *
-      * @throws java.lang.RuntimeException if the condition is false
-      * @group Boolean Type
-      * @see [[org.apache.spark.sql.functions.assert_true(c:org\.apache\.spark\.sql\.Column):* org.apache.spark.sql.functions.assert_true]]
-      */
-    def assertTrue: NullColumn = column.elem.map(f.assert_true).toDC
-
-    /**
-      * Returns null if the condition is true; throws an exception with the error message otherwise.
-      *
-      * @throws java.lang.RuntimeException if the condition is false
-      * @group Boolean Type
-      * @see [[org.apache.spark.sql.functions.assert_true(c:org\.apache\.spark\.sql\.Column,e:* org.apache.spark.sql.functions.assert_true]]
-      */
-    def assertTrue(msg: StringColumn): NullColumn =
-      (column.elem, msg.elem).mapN(f.assert_true).toDC
   }
 }
