+++ core/src/test/scala/doric/syntax/TypeMatcherSpec.scala
@@ -63,7 +63,9 @@ class TypeMatcherSpec
 
         val errors = testColumn.elem.run(df).toEither.left.value
         errors.length shouldBe 1
-        errors.head.message shouldBe "Cannot resolve column name \"int2\" among (colArr, int, str)"
+        errors.head.message should startWith(
+          "Cannot resolve column name \"int2\" among (colArr, int, str)"
+        )
       }
 
       it(
@@ -76,7 +78,9 @@ class TypeMatcherSpec
 
         val errors = testColumn.elem.run(df).toEither.left.value
         errors.length shouldBe 1
-        errors.head.message shouldBe "Cannot resolve column name \"int3\" among (colArr, int, str)"
+        errors.head.message should startWith(
+          "Cannot resolve column name \"int3\" among (colArr, int, str)"
+        )
       }
 
       it(
